nogit/aws/requirements-dev.txt
boto3==1.18.15
Faker==8.0.0 
qrcode==7.2
Pillow==8.3.1
marshmallow==3.7.1
requests==2.25.1
python-jose==3.2.0
pytz==2021.3
redis==3.5.3
PyJWT==1.7.1
python-jose==3.2.0
aws-encryption-sdk == 3.1.0
slack-sdk == 3.16.0
pandas-profiling == 3.2.0
user-agents == 2.2.0
kafka-python == 2.0.1
phonenumbers==8.13.1
numpy == 1.24.2
pandas == 1.5.3
